Create Queue Optimization.py
There are n people at the Railway Station who are to buy tickets to go to k different destinations. There are m ticket windows from which the ticket can be purchased. Each of the ticketing counter has an offer. If a person in the queue shares the same destination as the person in front of him/her, a 20% reduction in the ticket price is offered to him/her i.e.,

Say there are 3 people in the queue all with the same destination which costs 10 bucks, the first person in the queue pays 10 bucks. The 2nd and 3rd person get a discount of 20% on 10 bucks. So, they end up paying 8 bucks each instead of 10 bucks.

Try to distribute n people across m counters such that the total cost S paid by all n people is minimized.

Say if the queue has 5 people, 1,2,3,4,5. During distributing people among various queues, 1 goes before 2 and 2 goes before 3. The order is maintained.